feat: add recovery service and refactor the related tasks
